perf: Performance Optimization (#12)
* Refactor EmitWeatherParticleSystem

* Share a single particle pool among all emitters.

* Turn the particle emitters into children of the player entity.

This prevents having to recreate them whenever
the player is moving.

* Stop persisting particle emitters, instead recreate them.

There is no point in saving particle emitter
data. They get recreated as soon as the player
spawns or connects.

* Fix setting the weather via command line.

* Distribute particle evenly over the particle area.

The amount of particle emitters is now independent
of the size of the particle area. It can be
explicitly specified.

* Tweak weather related config values.

* Use Name instances to represent weather instead of strings.

* Use a gaussian distribution to determine the emitter positions.

This makes the particle effect more dense
in the field of view of the player.

* Create particles only for the local player

* Delay the particle effect until LocalPlayer has been initialized.

* Fix relative positioning of emitters.